Removes unnecessary service interfaces
