+++ docs/edu/hawaii/its/api/service/GroupAttributeServiceImpl.html
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>edu.hawaii.its.api.service.GroupAttributeServiceImpl</li>
+<li>edu.hawaii.its.api.service.GroupAttributeService</li>
 </ul>
 </li>
 </ul>
