+++ docs/edu/hawaii/its/api/service/GroupingAssignmentServiceImpl.html
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>edu.hawaii.its.api.service.GroupingAssignmentServiceImpl</li>
+<li>edu.hawaii.its.api.service.GroupingAssignmentService</li>
 </ul>
 </li>
 </ul>
