+++ docs/edu/hawaii/its/api/service/MemberAttributeServiceImpl.html
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>edu.hawaii.its.api.service.MemberAttributeServiceImpl</li>
+<li>edu.hawaii.its.api.service.MemberAttributeService</li>
 </ul>
 </li>
 </ul>
