+++ docs/edu/hawaii/its/api/service/MembershipServiceImpl.html
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>edu.hawaii.its.api.service.MembershipServiceImpl</li>
+<li>edu.hawaii.its.api.service.MembershipService</li>
 </ul>
 </li>
 </ul>
