+++ src/main/java/edu/hawaii/its/api/service/GroupingsMailService.java
@@ -1,29 +1,129 @@
 package edu.hawaii.its.api.service;
 
+import com.opencsv.CSVWriter;
 import edu.hawaii.its.api.type.UIAddMemberResults;
-import org.springframework.mail.javamail.JavaMailSender;
 
+import org.springframework.core.io.FileSystemResource;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.mail.javamail.MimeMessageHelper;
+import org.springframework.stereotype.Service;
+
+import javax.mail.MessagingException;
+import javax.mail.internet.MimeMessage;
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.util.ArrayList;
 import java.util.List;
 
 /**
  * Create and send SMTP messages using the JavaMailSender.
  */
-public interface GroupingsMailService {
-    GroupingsMailService setJavaMailSender(JavaMailSender javaMailSender);
+@Service("groupingsMailService")
+public class GroupingsMailService {
+    JavaMailSender javaMailSender;
+    String from = "";
 
-    GroupingsMailService setFrom(String from);
+    public GroupingsMailService setJavaMailSender(JavaMailSender javaMailSender) {
+        this.javaMailSender = javaMailSender;
+        return this;
+    }
+
+    public GroupingsMailService setFrom(String from) {
+        this.from = from;
+        return this;
+    }
 
     /**
-     * Send a SMTP message with no attachment
+     * Send a SMTP message with no attachment.
      */
-    void sendSimpleMessage(String from, String to, String subject, String text);
+    public void sendSimpleMessage(String from, String to, String subject, String text) {
 
+        SimpleMailMessage simpleMailMessage = new SimpleMailMessage();
+        simpleMailMessage.setFrom(from);
+        simpleMailMessage.setTo(to);
+        simpleMailMessage.setSubject(subject);
+        simpleMailMessage.setText(text);
+        this.javaMailSender.send(simpleMailMessage);
 
-    void sendCSVMessage(String from, String to, String subject, String text, String path,
-            List<UIAddMemberResults> res);
+    }
 
     /**
-     * Concat UH email suffix onto username
+     * Send an SMTP message with a CSV file attachment.
      */
-    String getUserEmail(String uid);
+    public void sendCSVMessage(String from, String to, String subject, String text, String path,
+            List<UIAddMemberResults> res) {
+
+        File file = new File(path);
+
+        try {
+            this.sendAttachmentMessage(from, to, subject, text, this.toCsv(this.toCsvObj(res), file), file);
+        } catch (IOException ioe) {
+            ioe.printStackTrace();
+        }
+    }
+
+    /**
+     * Send an SMTP message with a file attachment
+     */
+    private void sendAttachmentMessage(String from, String to, String subject, String text,
+            FileSystemResource fileSystemResource, File file) {
+
+        try {
+            MimeMessage mimeMessage = this.javaMailSender.createMimeMessage();
+            MimeMessageHelper mimeMessageHelper = new MimeMessageHelper(mimeMessage, true);
+
+            mimeMessageHelper.setFrom(from);
+            mimeMessageHelper.setTo(to);
+            mimeMessageHelper.setSubject(subject);
+            mimeMessageHelper.setText(text);
+
+            mimeMessageHelper.addAttachment(file.getPath(), fileSystemResource);
+            javaMailSender.send(mimeMessage);
+            file.delete();
+
+        } catch (MessagingException me) {
+            me.printStackTrace();
+        }
+
+    }
+
+    /**
+     * Build the lines of a CSV file.
+     */
+    private List<String[]> toCsvObj(List<UIAddMemberResults> addMemberResults) {
+        List<String[]> lines = new ArrayList<>();
+
+        lines.add(new String[] { "uid", "uhUuid", "name", "result" });
+
+        for (UIAddMemberResults addMemberResult : addMemberResults) {
+            lines.add(addMemberResult.toCsv());
+        }
+        return lines;
+    }
+
+    /**
+     * Write Csv data to a file.
+     */
+    private FileSystemResource toCsv(List<String[]> data, File file) throws IOException {
+        try {
+            FileWriter out = new FileWriter(file);
+            CSVWriter writer = new CSVWriter(out);
+
+            writer.writeAll(data);
+            writer.close();
+
+        } catch (FileNotFoundException e) {
+            throw new IOException(e);
+        }
+
+        return new FileSystemResource(file);
+    }
+
+    /* Concat UH suffix onto username */
+    public String getUserEmail(String uid) {
+        return uid + "@hawaii.edu";
+    }
 }
