+++ src/test/java/edu/hawaii/its/api/service/TestGrouperApiService.java
@@ -255,7 +255,7 @@ public class TestGrouperApiService {
 
         try {
             grouperApiService.hasMemberResults(GROUPING_INCLUDE, new Person(null, null, null)).getResults();
-        } catch (NullPointerException e) {
+        } catch (IllegalArgumentException e) {
             assertEquals("The person is required to have either a username or a uuid", e.getMessage());
         }
     }
