+++ src/test/java/edu/hawaii/its/api/wrapper/GroupsResultsTest.java
@@ -1,36 +1,30 @@
 package edu.hawaii.its.api.wrapper;
 
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Test;
-import edu.hawaii.its.api.util.JsonUtil;
-
-import edu.internet2.middleware.grouperClient.ws.beans.WsGetGroupsResults;
-
-import java.io.FileInputStream;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.List;
-import java.util.Properties;
-
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import edu.hawaii.its.api.util.JsonUtil;
+import edu.hawaii.its.api.util.PropertyLocator;
+import edu.internet2.middleware.grouperClient.ws.beans.WsGetGroupsResults;
+
+import java.util.List;
+
 public class GroupsResultsTest {
 
-    private static Properties properties;
+    private PropertyLocator propertyLocator;
 
-    @BeforeAll
-    public static void beforeAll() throws Exception {
-        Path path = Paths.get("src/test/resources");
-        Path file = path.resolve("grouper.test.properties");
-        properties = new Properties();
-        properties.load(new FileInputStream(file.toFile()));
+    @BeforeEach
+    public void beforeEach() throws Exception {
+        propertyLocator = new PropertyLocator("src/test/resources", "grouper.test.properties");
     }
 
     @Test
     public void construction() {
-        String json = propertyValue("groups.results");
+        String json = propertyLocator.find("groups.results");
         WsGetGroupsResults wsGetGroupsResults = JsonUtil.asObject(json, WsGetGroupsResults.class);
         GroupsResults groupsResults = new GroupsResults(wsGetGroupsResults);
         assertNotNull(wsGetGroupsResults);
@@ -42,7 +36,7 @@ public class GroupsResultsTest {
 
     @Test
     public void groupPathsTest() {
-        String json = propertyValue("groups.results");
+        String json = propertyLocator.find("groups.results");
         WsGetGroupsResults wsGetGroupsResults = JsonUtil.asObject(json, WsGetGroupsResults.class);
         GroupsResults groupsResults = new GroupsResults(wsGetGroupsResults);
         List<String> groupPaths = groupsResults.groupPaths();
@@ -56,7 +50,7 @@ public class GroupsResultsTest {
         assertNotNull(groupPaths);
         assertTrue(groupPaths.isEmpty());
 
-        json = propertyValue("groups.results.empty.results");
+        json = propertyLocator.find("groups.results.empty.results");
         wsGetGroupsResults = JsonUtil.asObject(json, WsGetGroupsResults.class);
         groupsResults = new GroupsResults(wsGetGroupsResults);
         groupPaths = groupsResults.groupPaths();
@@ -65,7 +59,7 @@ public class GroupsResultsTest {
         assertNotNull(groupPaths);
         assertTrue(groupPaths.isEmpty());
 
-        json = propertyValue("groups.results.empty.groups");
+        json = propertyLocator.find("groups.results.empty.groups");
         wsGetGroupsResults = JsonUtil.asObject(json, WsGetGroupsResults.class);
         groupsResults = new GroupsResults(wsGetGroupsResults);
         groupPaths = groupsResults.groupPaths();
@@ -73,9 +67,5 @@ public class GroupsResultsTest {
         assertNotNull(groupsResults);
         assertNotNull(groupPaths);
         assertTrue(groupPaths.isEmpty());
-    }
-
-    private String propertyValue(String key) {
-        return properties.getProperty(key);
     }
 }
